% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S7primitivereprex-package.R
\docType{package}
\name{S7primitivereprex-package}
\alias{S7primitivereprex}
\alias{S7primitivereprex-package}
\title{S7primitivereprex: What the Package Does}
\description{
The goal of S7primitivereprex is to demonstrate an S7 bug.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/S7primitivereprex}
  \item Report bugs at \url{https://github.com/jonthegeek/S7primitivereprex/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
